--- src/components/PopUpWindow.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import Popup from 'reactjs-popup';
-import 'reactjs-popup/dist/index.css';
-
-export default () => (
-    <Popup trigger={<button> Trigger</button>} position="right center">
-        <div>Popup content here !!</div>
-    </Popup>
-);
