+++ src/components/PopUpWindow/PopUpWindow.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import Popup from 'reactjs-popup';
+import 'reactjs-popup/dist/index.css';
+
+const PopUpWindow = () => (
+    <Popup trigger={<button> Trigger</button>} position="right center">
+        <div>Popup content here !!</div>
+    </Popup>
+);
+
+export default PopUpWindow;
